{"types":"C","runs":33}

+++ Problem16.java
@@ -2,9 +2,9 @@ import acm.program.ConsoleProgram;
 
 public class Problem16 extends ConsoleProgram {
 	public void run() {
-		int a = readInt("First number: ");
-		int b = readInt("Second number: ");
-		int c = readInt("Third number: ");
+		double a = readDouble("First number: ");
+		double b = readDouble("Second number: ");
+		double c = readDouble("Third number: ");
 		println(1 / (1/a + 1/b + 1/c));
 	}
 }
